# swagger.yml
swagger: "2.0"
info:
  description: A demo app for LAIN
  title: A todo list application
  version: 1.0.0
consumes:
  - application/vnd.laincloud.todomvc.v1+json
produces:
  - application/vnd.laincloud.todomvc.v1+json
schemes:
  - http
paths:
...
  /todos/{id}:
    parameters:
      - name: id
        in: path
        type: string
        required: true
    get:
      operationId: getTodo
      description: "Get a todo item"
      tags:
        - todo
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/todo"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/error"
    put:
      operationId: updateTodo
      description: "Update a todo item"
      tags:
        - todo
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/todo"
      responses:
        "204":
          description: "OK"
          schema:
            $ref: "#/definitions/todo"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/error"
    delete:
      operationId: deleteTodo
      description: "Delete a todo item"
      tags:
        - todo
      responses:
        "204":
          description: "Deleted"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/error"
definitions:
  error:
    type: object
    required:
      - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
  todo:
    type: object
    required:
      - title
      - done
    properties:
      id:
        type: integer
        format: int64
        readOnly: true
      title:
        type: string
        minLength: 1
      done:
        type: boolean